+++ go.mod
@@ -2,4 +2,4 @@ module github.com/Yandex-Practicum/go-rest-api-homework
 
 go 1.20
 
-require github.com/go-chi/chi/v5 v5.0.10 // indirect
+require github.com/go-chi/chi/v5 v5.0.10
