+++ pkg/task/task.go
@@ -1,0 +1,9 @@
+package task
+
+// Task ...
+type Task struct {
+	ID           string   `json:"id"`
+	Description  string   `json:"description"`
+	Note         string   `json:"note"`
+	Applications []string `json:"applications"`
+}
